+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+## [2.47.3] 2021-08-23
+
+- hardis:org:diagnose:legacyapi: Add more summary and statistics
+
 ## [2.47.2] 2021-08-23
 
 - Fix hardis:org:diagnose:legacyapi (display raw logs when CSV builder is crashing) , using papaparse instead of objects-to-csv package